% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxalook.R
\name{tl_query}
\alias{tl_query}
\title{Query the taxa lookup up database.}
\usage{
tl_query(
  tax = NULL,
  tax_match = "fuzzy",
  tl_id = NULL,
  taxonomy = FALSE,
  id = FALSE,
  group = FALSE,
  habitat = FALSE,
  continent = FALSE,
  country = FALSE
)
}
\arguments{
\item{tax}{character vector; Which taxa should be returned?}

\item{tax_match}{Should \code{tax} names be matched exactly or by pattern?
Can be one of fuzzy (default) or exact.}

\item{tl_id}{integer vector; Should a specific taxalook id entry be returned?}

\item{taxonomy}{logical; Should taxonomic information be returned?}

\item{id}{logical; Should identifiers for other databases be returned?}

\item{group}{logical; Should ecological groups be returned?}

\item{habitat}{logical; Should habitat information be returned?}

\item{continent}{logical; Should continental occurrence information be returned?}

\item{country}{logical; Should country occurrence information be returned?}
}
\value{
Returns a list of three data.tables (filtered data base query results, aggregated data base query results, meta information)
}
\description{
Query the taxa lookup up database.
}
\examples{
tl_query(tax = 'Oncorhynchus', group = TRUE)
tl_query(tax = c('Daphnia', 'Perla'), habitat = TRUE, tax_match = 'exact')
tl_query(tax = 'quercus', country = TRUE, id = TRUE)

}
\author{
Andreas Scharmueller \email{andschar@protonmail.com}
}
