% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxalook.R
\name{tl_query}
\alias{tl_query}
\title{Query the taxa lookup up database.}
\usage{
tl_query(
  query = NULL,
  query_match = "exact",
  from = NULL,
  what = "id",
  force_download = FALSE
)
}
\arguments{
\item{query}{character vector; Which taxa should be returned?}

\item{query_match}{Should \code{query} names be matched exactly or by pattern?
Can be one of exact (default) or fuzzy.}

\item{from}{Which column should be used for matching? Can be one of:
'tl_id', 'gbif_id', 'ncbi_id', 'worms_id' or 'epa_species_number'.
If NULL (default), queries are matched against all columns (for taxa names).}

\item{what}{What should be returned? Can be one or multiple of 'id' (default),
'continent', 'country', 'group', 'habitat', 'taxonomy' or 'trophic_lvl'.}

\item{force_download}{Force download anyway? Helpful if downloaded file is corrupt.}
}
\value{
Returns a data.table.
}
\description{
Query the taxa lookup up database.
}
\examples{
tl_query(query = 'Oncorhynchus', what = 'group')
tl_query(query = c('Daphnia', 'Perla'), what = 'habitat', query_match = 'exact')
tl_query(query = 'quercus', query_match = 'fuzzy', what = c('continent'))

}
\author{
Andreas Scharmueller \email{andschar@proton.me}
}
